requirements.txt
fastapi==0.75.0
uvicorn==0.17.6
uvloop==0.16.0
PyCryptodome==3.10.1
eciespy==0.3.11
